requirements.txt
boto3
azure-mgmt-security
azure-identity
google-cloud-securitycenter
google-auth
requests 